Part 1 Completion

+++ AoC/01/aoc-01.h
@@ -12,7 +12,7 @@ namespace AoC {
 		EDirection_MAX
 	};
 
-	typedef std::pair<EDirection, uint8_t> instruction; 
+	typedef std::pair<EDirection, uint16_t> instruction; 
 	
 	std::vector<instruction> ParseInstructions(const std::string& instructions);
 	uint64_t PickLock(const uint8_t startingNum, const std::vector<instruction>& instructions);
